+++ lisp/eab-org-agenda.el
@@ -3,8 +3,8 @@
 (defun eab/replace-in-string (what with in)
   (replace-regexp-in-string (regexp-quote what) with in nil 'literal))
 
-(setq eab/org-ql-H-query '(and (or (not (tags "noagenda")) (tags "agenda")) (clocked 560) ))
-(setq eab/org-ql-O-query '(and (or (not (tags "noagenda")) (tags "agenda")) (clocked 6000) (not (clocked 1000))))
+(setq eab/org-ql-H-query '(and (or (not (tags "noagenda")) (tags "agenda")) (clocked 600) ))
+(setq eab/org-ql-O-query '(and (or (not (tags "noagenda")) (tags "agenda")) (clocked 6000) (not (clocked 600))))
 (setq eab/org-ql-W-query '(and (and (or (tags "w1c") (tags "fz")) (or (not (tags "noagenda")) (tags "agenda"))) (clocked 560)))